refactor: create three column layout
